% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_time.R
\name{combine_time}
\alias{combine_time}
\title{Combine Close Observations Based on a Tolerance Threshold}
\usage{
combine_time(time, tol)
}
\arguments{
\item{time}{A numeric vector containing the observations to be adjusted.}

\item{tol}{A numeric value representing the tolerance threshold, defined as
a fraction of the difference between the first and last values in \code{time}.}
}
\value{
A modified numeric vector where values within the tolerance
threshold have been merged with the preceding value.
}
\description{
This function modifies a numeric vector of observations by merging values
that are within a specified tolerance threshold.
}
\references{
This function was extracted from the GitHub repository
\href{https://github.com/zjph602xtc/PWEXP}{PWEXP}.
}
\author{
Tianchen Xu (\href{mailto:zjph602xutianchen@gmail.com}{zjph602xutianchen@gmail.com})
}
