% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pwpldis.R
\name{fit_pwpldis}
\alias{fit_pwpldis}
\title{Fit the Discrete Piecewise Power-law Distribution}
\usage{
fit_pwpldis(
  time,
  breakpoint = NULL,
  nbreak = NULL,
  exclude_int = NULL,
  min_pt_tail = 2,
  max_set = 10000,
  trace = FALSE,
  tol = 1e-04
)
}
\arguments{
\item{time}{A numeric vector of observations.}

\item{breakpoint}{A numeric vector specifying fixed change points.}

\item{nbreak}{An integer specifying the total number of change points in the
model. This includes any change points specified in the \code{breakpoint} argument.}

\item{exclude_int}{A numeric vector of two values defining an interval in which
estimated change points should be excluded (e.g., \code{exclude_int = c(10, Inf)}
excludes change points after \code{time = 10}).}

\item{min_pt_tail}{An integer specifying the minimum number of observations
required to estimate the last parameter.}

\item{max_set}{An integer specifying the maximum number of possible breakpoint
combinations to consider.}

\item{trace}{A logical value indicating whether to display the optimization
process. Default is \code{FALSE}.}

\item{tol}{A numeric value specifying the minimum allowed gap between two
change points.}
}
\value{
A data frame with the following columns:
\describe{
\item{taus}{Estimated change points.}
\item{alphas}{Estimated scaling parameters for each piece.}
\item{likelihood}{The log-likelihood of the model.}
\item{AIC}{The Akaike Information Criterion for the model.}
\item{BIC}{The Bayesian Information Criterion for the model.}
}
The result also contains two additional attributes:
\describe{
\item{brk}{A vector of the estimated change points.}
\item{lam}{A vector of the estimated scaling parameters.}
}
}
\description{
This function fits a piecewise power-law distribution to datasets. The user
can specify all change points, some change points, or allow the function to
estimate them. It also provides the option to exclude intervals and specify
the minimum number of observations required to estimate the last parameter.
}
\details{
If the user specifies change points, the function checks their validity and removes
change points after the last event time or before the first event time. If \code{nbreak} is
set to 0, the function estimates the number of change points based on the number of
unique observations. The \code{exclude_int} argument allows excluding change points that are
too close to the tail. The \code{min_pt_tail} argument helps in estimating a more robust
scaling parameter for the tail by requiring a minimum number of observations.
}
\examples{
# Initial specification
set.seed(2025)
p <- c(1, 3, 5)
alpha <- c(1.5, 2, 3)

# Generate a dataset from the discrete piecewise power-law model
df <- rpwpldis(300, p, alpha)

# Fit the model with fixed change points (cps) at 3 and 5
fit_pwpldis(df, breakpoint = c(3, 5))

# Fit the model allowing it to estimate 1 cp
fit_pwpldis(df, nbreak = 1)

# Fit the model with 1 estimated cp and display optimization process
fit_pwpldis(df, nbreak = 1, trace = TRUE)

# Fit the model with 1 estimated cp, requiring at least 100 observations in the last partition
fit_pwpldis(df, nbreak = 1, min_pt_tail = 100)

# Compute the number of observations in each interval defined by cps
n_each_interval(df, c(1, 4))

# Fit the model allowing it to estimate 2 cps
fit_pwpldis(df, nbreak = 2)

# Fit the model with 2 estimated cps, excluding cps beyond 5
fit_pwpldis(df, nbreak = 2, exclude_int = c(5, Inf))

}
\references{
This function was inspired by the implementation of the \code{PWEXP}
package in \code{R}. Visit their GitHub repository
\href{https://github.com/zjph602xtc/PWEXP}{PWEXP}.
}
