% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid.R
\name{get_grid}
\alias{get_grid}
\title{Generate a Grid of Breakpoints}
\usage{
get_grid(time, breakpoint, nbreak, max_set = 5000, remove_first = TRUE)
}
\arguments{
\item{time}{A numeric vector of observations.}

\item{breakpoint}{A numeric vector specifying fixed change points.}

\item{nbreak}{An integer specifying the number of change points in the model.}

\item{max_set}{An integer specifying the maximum number of combinations to generate.
Default is 5000.}

\item{remove_first}{A logical value indicating whether to exclude the first time point.
Default is \code{TRUE}.}
}
\value{
A matrix where each row represents a combination of observations
forming a set of change points for the model.
}
\description{
This function constructs a grid of change points for the discrete piecewise
power-law model. It selects a subset of unique observations and ensures that
the number of combinations does not exceed a specified limit.
}
\references{
This function was extracted from the GitHub repository
\href{https://github.com/zjph602xtc/PWEXP}{PWEXP}.
}
\author{
Tianchen Xu \href{mailto:zjph602xutianchen@gmail.com}{zjph602xutianchen@gmail.com}
}
