% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/each_interval.R
\name{index_each_interval}
\alias{index_each_interval}
\alias{n_each_interval}
\title{Observation Indexing and Counting in Defined Partitions}
\usage{
index_each_interval(x, p)

n_each_interval(x, p)
}
\arguments{
\item{x}{A numeric vector of observations.}

\item{p}{A numeric vector specifying the minimum value \eqn{\tau_{(0)}} and
change points \eqn{\tau_{(1)}, \ldots, \tau_{(k)}} that define the partitions,
i.e., \eqn{p = (\tau_{(0)}, \tau_{(1)}, \ldots, \tau_{(k)})}. It must be sorted
in increasing order.}
}
\value{
\itemize{
\item \code{index_each_interval()}: A list of integer vectors, where each element
contains the indices of observations belonging to the corresponding interval.
\item \code{n_each_interval()}: A numeric vector indicating the count of observations
in each interval.
}
}
\description{
These functions identify and count the number of observations within
specified intervals defined by the change points in \code{p}. The intervals are
assumed to be left-closed and right-open.
}
\details{
The intervals follow the structure:
\deqn{\mathcal{R}_1 = [\tau_{(0)}, \tau_{(1)}), \mathcal{R}_2 = [\tau_{(1)}, \tau_{(2)}),
..., \mathcal{R}_{k} = [\tau_{(k-1)}, \tau_{(k)}),}
except for the last partition, which is closed: \eqn{\mathcal{R}_{k+1} = [\tau_{(k)}, +\infty)}.

The function assumes that \code{p} and \code{alpha} have the same length.
}
\examples{
# Example data
x <- c(7, 7, 6, 4, 1, 5, 2, 6, 3, 5)

# Get indices of observations within each interval
index_each_interval(x, c(1, 3, 5))

# Count observations in each interval
n_each_interval(x, c(1, 3, 5))

}
