% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwpldis.R
\name{dpwpldis}
\alias{dpwpldis}
\alias{ppwpldis}
\alias{spwpldis}
\alias{hpwpldis}
\alias{qpwpldis}
\alias{rpwpldis}
\title{Discrete Piecewise Power-Law Distribution}
\usage{
dpwpldis(x, p, alpha)

ppwpldis(x, p, alpha)

spwpldis(x, p, alpha)

hpwpldis(x, p, alpha)

qpwpldis(u, p, alpha)

rpwpldis(n, p, alpha)
}
\arguments{
\item{x}{A numeric value to evaluate the function.}

\item{p}{A numeric vector specifying the minimum value \eqn{\tau_{(0)}} and
change points \eqn{\tau_{(1)}, \ldots, \tau_{(k)}} that define the partitions,
i.e., \eqn{p = (\tau_{(0)}, \tau_{(1)}, \ldots, \tau_{(k)})}. It must be sorted
in increasing order.}

\item{alpha}{A numeric vector of scaling parameters
\eqn{\alpha_{(1)}, \ldots, \alpha_{(k + 1)}} corresponding to each partition.
All values must be greater than 1.}

\item{u}{A probability to evaluate the quantile function.}

\item{n}{An integer specifying the number of random samples to generate.}
}
\value{
\itemize{
\item \code{dpwpldis()}: The probability mass function (PMF) evaluated at \code{x}.
\item \code{ppwpldis()}: The cumulative distribution function (CDF) evaluated at \code{x}.
\item \code{spwpldis()}: The survival function (1 - CDF) evaluated at \code{x}.
\item \code{hpwpldis()}: The hazard function evaluated at \code{x}.
\item \code{qpwpldis()}: The quantile function, returning the smallest \code{x} such that
\eqn{P(X \leq x) \geq u}
\item \code{rpwpldis()}: A numeric vector of \code{n} random samples drawn from the distribution.
}
}
\description{
This set of functions provides the density, cumulative distribution function,
survival function, hazard function, quantile function, and random generation
for the discrete piecewise power-law distribution.
}
\details{
The function assumes that \code{p} and \code{alpha} have the same length.
Some functions rely on the Hurwitz zeta function from the
VGAM package.
}
\examples{
p <- c(1, 3, 5)
alpha <- c(1.5, 2, 3)
dpwpldis(5, p, alpha)
ppwpldis(5, p, alpha)
spwpldis(5, p, alpha)
hpwpldis(5, p, alpha)
qpwpldis(0.5, p, alpha)
rpwpldis(10, p, alpha)

}
