% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_pwpldis.R
\name{loglik_pwpldis}
\alias{loglik_pwpldis}
\title{Log-Likelihood of the Discrete Piecewise Power-Law Model}
\usage{
loglik_pwpldis(alpha, x, p)
}
\arguments{
\item{alpha}{A numeric vector of scaling parameters
\eqn{\alpha_{(1)}, \ldots, \alpha_{(k + 1)}} corresponding to each partition.
All values must be greater than 1.}

\item{x}{A numeric vector of observed data values.}

\item{p}{A numeric vector specifying the minimum value \eqn{\tau_{(0)}} and
change points \eqn{\tau_{(1)}, \ldots, \tau_{(k)}} that define the partitions,
i.e., \eqn{p = (\tau_{(0)}, \tau_{(1)}, \ldots, \tau_{(k)})}. It must be sorted
in increasing order.}
}
\value{
A numeric value representing the log-likelihood of the data given
the specified parameters and change points
}
\description{
This function computes the log-likelihood for a discrete piecewise power-law model.
It evaluates the log-likelihood given a set of parameters, observed data, and
the change points that define the intervals.
}
\details{
The function assumes that \code{p} and \code{alpha} have the same length.
The calculation involves the Hurwitz zeta function from the
VGAM package.
}
\examples{
x <- c(7, 7, 6, 4, 1, 5, 2, 6, 3, 5)
loglik_pwpldis(c(2.0, 3.5), x, c(1, 4))
loglik_pwpldis(c(2.5, 3.0), x, c(1, 4))
loglik_pwpldis(c(1.5, 2.0), x, c(1, 4))

}
